Added new class UserManagementConfiguration and changed name of the class SecurityConfiguration to WebAuthorizationConfiguration to separate responsibilities
