+++ javadoc/pl/pjatk/MATLOG/UserManagement/SecurityConfiguration.html
@@ -83,7 +83,7 @@ loadScripts(document, 'script');</script>
 <h1 title="Class SecurityConfiguration" class="title">Class SecurityConfiguration</h1>
 </div>
 <div class="inheritance" title="Inheritance Tree"><a href="https://docs.oracle.com/en/java/javase/23/docs/api/java.base/java/lang/Object.html" title="class or interface in java.lang" class="external-link">java.lang.Object</a>
-<div class="inheritance">pl.pjatk.MATLOG.userManagement.SecurityConfiguration</div>
+<div class="inheritance">pl.pjatk.MATLOG.userManagement.WebAuthorizationConfiguration</div>
 </div>
 <section class="class-description" id="class-description">
 <div class="horizontal-scroll">
