+++ src/main/java/pl/pjatk/MATLOG/userManagement/UserManagementConfiguration.java
@@ -1,0 +1,26 @@
+package pl.pjatk.MATLOG.userManagement;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.security.core.userdetails.UserDetailsService;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
+
+@Configuration
+public class UserManagementConfiguration {
+
+    @Bean
+    public UserDetailsService userDetailsService(UserService userService) {
+        return new MongoUserDetailsService(userService);
+    }
+
+    @Bean
+    public PasswordEncoder passwordEncoder() {
+        return new BCryptPasswordEncoder();
+    }
+
+    @Bean
+    public UserPasswordValidator passwordValidator() {
+        return new StandardUserPasswordValidator();
+    }
+}
