+++ src/main/java/pl/pjatk/MATLOG/userManagement/WebAuthorizationConfiguration.java
@@ -4,14 +4,11 @@ import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.SecurityFilterChain;
 
 @Configuration
 @EnableWebSecurity
-public class SecurityConfiguration {
+public class WebAuthorizationConfiguration {
 
     @Bean
     public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
@@ -26,20 +23,5 @@ public class SecurityConfiguration {
                             .anyRequest().authenticated()
                 );
         return http.build();
-    }
-
-    @Bean
-    public UserDetailsService userDetailsService(UserService userService) {
-        return new MongoUserDetailsService(userService);
-    }
-
-    @Bean
-    public PasswordEncoder passwordEncoder() {
-        return new BCryptPasswordEncoder();
-    }
-
-    @Bean
-    public UserPasswordValidator passwordValidator() {
-        return new StandardUserPasswordValidator();
     }
 }
